# Reverts change around module name allocation