--- a/handler/middleware.go
+++ b/handler/middleware.go
@@ -8,6 +8,7 @@ import (
 	"strconv"
 	"strings"
 	"sync"
+	"sync/atomic"
 
 	"github.com/tetratelabs/wazero"
 	wazeroapi "github.com/tetratelabs/wazero/api"
@@ -52,6 +53,7 @@ type middleware struct {
 	logger                  api.Logger
 	pool                    sync.Pool
 	features                handler.Features
+	instanceCounter         uint64
 }
 
 func (m *middleware) Features() handler.Features {
@@ -74,10 +76,9 @@ func NewMiddleware(ctx context.Context, guest []byte, host handler.Host, opts ..
 	}
 
 	m := &middleware{
-		host:    host,
-		runtime: wr,
-		// Clear the module name, to avoid conflicts when the pool size is >1.
-		moduleConfig: o.moduleConfig.WithName(""),
+		host:         host,
+		runtime:      wr,
+		moduleConfig: o.moduleConfig,
 		guestConfig:  o.guestConfig,
 		logger:       o.logger,
 	}
@@ -197,7 +198,9 @@ type guest struct {
 }
 
 func (m *middleware) newGuest(ctx context.Context) (*guest, error) {
-	g, err := m.runtime.InstantiateModule(ctx, m.guestModule, m.moduleConfig)
+	moduleName := fmt.Sprintf("%d", atomic.AddUint64(&m.instanceCounter, 1))
+
+	g, err := m.runtime.InstantiateModule(ctx, m.guestModule, m.moduleConfig.WithName(moduleName))
 	if err != nil {
 		_ = m.runtime.Close(ctx)
 		return nil, fmt.Errorf("wasm: error instantiating guest: %w", err)