This reverts to the former approach of external name management because it ends up faster in practice with http-wasm.

The last commit used "anonymous modules" which is not a good fit for http-wasm in how it works today. Basically, http-wasm is pooling modules and these are re-used for sequential handler functions. The members of the pool change at a very low rate, so there's not a lot of lock contention on the module namespace. On the other hand, the normal path, sequential performance is slower when we use anonymous modules.

```
$ benchstat old.txt new.txt 
goos: darwin
goarch: arm64
pkg: github.com/http-wasm/http-wasm-host-go/handler/nethttp
                                                                │   old.txt   │              new.txt               │
                                                                │   sec/op    │   sec/op     vs base               │
/read_body/read_body-12                                           505.9n ± 4%   471.7n ± 2%   -6.77% (p=0.002 n=6)
/get_header_names_none/get_header_names_none-12                   554.4n ± 2%   529.9n ± 2%   -4.40% (p=0.002 n=6)
/get_uri/get_uri-12                                               474.6n ± 3%   445.6n ± 0%   -6.12% (p=0.002 n=6)
/set_uri/set_uri-12                                               578.0n ± 2%   539.9n ± 1%   -6.59% (p=0.002 n=6)
/get_header_names/get_header_names-12                             538.2n ± 1%   508.1n ± 1%   -5.59% (p=0.002 n=6)
/get_header_values_exists/get_header_values_exists-12             485.1n ± 2%   450.2n ± 1%   -7.21% (p=0.002 n=6)
/get_header_values_not_exists/get_header_values_not_exists-12     510.3n ± 3%   471.1n ± 1%   -7.67% (p=0.002 n=6)
/read_body_stream/read_body_stream-12                             503.7n ± 2%   470.8n ± 0%   -6.53% (p=0.002 n=6)
/read_body_stream_large/read_body_stream_large-12                 625.8n ± 1%   599.0n ± 3%   -4.27% (p=0.002 n=6)
/example_wasi/example_wasi-12                                     3.348µ ± 1%   3.219µ ± 4%   -3.84% (p=0.015 n=6)
/example_router_host_response/example_router_host_response-12     565.6n ± 1%   533.0n ± 0%   -5.76% (p=0.002 n=6)
/log/log-12                                                       454.0n ± 4%   415.9n ± 1%   -8.38% (p=0.002 n=6)
/add_header_value/add_header_value-12                             626.6n ± 2%   573.3n ± 1%   -8.49% (p=0.002 n=6)
/set_status_code/set_status_code-12                               463.0n ± 1%   428.6n ± 2%   -7.43% (p=0.002 n=6)
/write_body/write_body-12                                         479.6n ± 2%   449.9n ± 1%   -6.19% (p=0.002 n=6)
/example_router_guest_response/example_router_guest_response-12   656.6n ± 2%   605.7n ± 2%   -7.76% (p=0.002 n=6)
/set_header_value/set_header_value-12                             633.3n ± 3%   577.0n ± 3%   -8.90% (p=0.002 n=6)
/remove_header/remove_header-12                                   521.6n ± 3%   489.2n ± 1%   -6.22% (p=0.002 n=6)
/get_header_names_large/get_header_names_large-12                 1.348µ ± 8%   1.043µ ± 1%  -22.60% (p=0.002 n=6)
geomean                                                           619.8n        573.3n
```